Add Storyteller transcript ingest and anchored alignment to forge pipeline
Forge matches now ingest Storyteller wordTimeline transcripts and run
align_storyteller_and_store (the same anchored N-Gram alignment used in
normal matches) to prevent 4-5 page drift. Also adds a local file
fallback when the Storyteller API download returns 404.

- Extract transcript ingestion helpers from web_server.py into
  alignment_service.py for shared use by both match and forge flows
- After forge completion, ingest chapter JSON files and set
  book.transcript_file / book.transcript_source
- Run Storyteller-anchored alignment with SMIL as fallback
- Add shutil.copy2 fallback in forge download when API is unavailable
